requirements.txt
Django~=5.1.4
mysqlclient==2.1.1
ffmpeg-python~=0.2.0
opencv-python~=4.10.0.84
numpy~=2.2.1
djangorestframework~=3.15.2
StrEnum~=0.4.15